requirements.txt
streamlit
requests
pydantic
sqlmodel
sqlalchemy
pillow
python-dotenv
